# Fix the passing of Javascript typed arrays to Python