Fixes a bug when passing typed array buffers from Javascript to Python (mentioned in https://github.com/iodide-project/pyodide/issues/59#issuecomment-400600336)

Cc: @gabrielfriere