@gabrielfreire 2018-06-27
This fix was huge. Now using `np.frombuffer()` 
- On my desktop (very fast hardware) i'm getting between **0.001s to 0.02s** to extract a spectrogram from an audio buffer
- On my work laptop (slower hardware) i'm getting between **0.07s to 0.7s** to extract a spectrogram from an audio buffer
this is pretty much native performance, great stuff!! really good.